+++ config/routes.oas.json
@@ -5,7 +5,7 @@
     "title": "My Zuplo API"
   },
   "paths": {
-    "/deploy": {
+    "/v1/deploy": {
       "x-zuplo-path": {
         "pathMode": "open-api"
       },
@@ -16,7 +16,7 @@
           "corsPolicy": "none",
           "handler": {
             "export": "default",
-            "module": "$import(./modules/deploy)",
+            "module": "$import(./v1/modules/deploy)",
             "options": {}
           },
           "policies": {
@@ -29,7 +29,7 @@
         "operationId": "e63fabdc-d80b-4ba8-9802-ba11426cf4bd"
       }
     },
-    "/contractAddress": {
+    "/v1/contractAddress": {
       "x-zuplo-path": {
         "pathMode": "open-api"
       },
@@ -40,7 +40,7 @@
           "corsPolicy": "none",
           "handler": {
             "export": "default",
-            "module": "$import(./modules/contractAddress)",
+            "module": "$import(./v1/modules/contractAddress)",
             "options": {}
           },
           "policies": {
@@ -53,7 +53,7 @@
         "operationId": "38437fbd-acd6-407b-8992-d4581ae061bd"
       }
     },
-    "/requestOwnership": {
+    "/v1/requestOwnership": {
       "x-zuplo-path": {
         "pathMode": "open-api"
       },
@@ -64,7 +64,7 @@
           "corsPolicy": "none",
           "handler": {
             "export": "default",
-            "module": "$import(./modules/requestOwnership)",
+            "module": "$import(./v1/modules/requestOwnership)",
             "options": {}
           },
           "policies": {
@@ -77,7 +77,7 @@
         "operationId": "e6df59eb-1345-4f89-be89-5a3435885734"
       }
     },
-    "/requestSignerChange": {
+    "/v1/requestSignerChange": {
       "x-zuplo-path": {
         "pathMode": "open-api"
       },
@@ -88,7 +88,7 @@
           "corsPolicy": "none",
           "handler": {
             "export": "default",
-            "module": "$import(./modules/requestSignerChange)",
+            "module": "$import(./v1/modules/requestSignerChange)",
             "options": {}
           },
           "policies": {
@@ -101,7 +101,7 @@
         "operationId": "0d212cee-1093-4c14-b423-74f7a5cadc71"
       }
     },
-    "/submitTransaction": {
+    "/v1/submitTransaction": {
       "x-zuplo-path": {
         "pathMode": "open-api"
       },
@@ -112,7 +112,7 @@
           "corsPolicy": "none",
           "handler": {
             "export": "default",
-            "module": "$import(./modules/submitTransaction)",
+            "module": "$import(./v1/modules/submitTransaction)",
             "options": {}
           },
           "policies": {
@@ -125,7 +125,7 @@
         "operationId": "22f21f24-8cfe-4e4b-b5df-0c11d271b623"
       }
     },
-    "/get-multisig": {
+    "/v1/get-multisig": {
       "x-zuplo-path": {
         "pathMode": "open-api"
       },
@@ -136,7 +136,7 @@
           "corsPolicy": "none",
           "handler": {
             "export": "default",
-            "module": "$import(./modules/getMultisig)",
+            "module": "$import(./v1/modules/getMultisig)",
             "options": {}
           },
           "policies": {
