+++ modules/v1/contractAddress.ts
@@ -1,8 +1,8 @@
 import { ethers } from "ethers";
 import { ZuploContext, ZuploRequest, environment } from "@zuplo/runtime";
 import { createClient } from "@supabase/supabase-js";
-import verifyNetwork from "./verification/verifyNetwork";
-import getRpcURL from "./verification/getRpcURL";
+import verifyNetwork from "../verification/verifyNetwork";
+import getRpcURL from "../verification/getRpcURL";
 
 const { SUPABASE_URL, SUPABASE_PASSWORD } = environment;
 
@@ -33,33 +33,6 @@ export default async function (request: ZuploRequest, context: ZuploContext) {
     }
   }
 
-  try {
-    const { data, error } = await supabase
-      .from("contracts")
-      .select("creation_hash")
-      .eq("creation_hash", txhash);
-    if (error) {
-      return {
-        statusCode: 503,
-        smallError: "Error in the connection with the database",
-        error: error,
-      };
-    }
-
-    if (!data || data.length == 0) {
-      return {
-        statusCode: 404,
-        error: "The transaction hash doesn't correspond to a MultiSigWallet deployed with the API"
-      }
-    }
-  } catch (err) {
-    return {
-      statusCode: 503,
-      smallError: "Error in the connection with the database",
-      error: err,
-    };
-  }
-
   const network = await request.query.network
   await verifyNetwork(network);
   const rpcUrl = await getRpcURL(network)
@@ -86,7 +59,6 @@ export default async function (request: ZuploRequest, context: ZuploContext) {
         if (error) {
           return {
             statusCode: 503,
-            smallError: "Error while inserting the contract's data in the database",
             error: error,
           };
         }
